Moving tests to respective files (#3)
* Moving tests to respective files

* Correcting typo

* Correcting typo

* Creating ReplayData and Fixing tests, updating Organization.py to take secret_type on secrets call
