+++ tests/ReplayData/Organization.testOrgGetSecretAssertion.txt
@@ -1,0 +1,10 @@
+https
+GET
+api.github.com
+None
+/orgs/demoorg
+{'Authorization': 'Basic login_and_password_removed', 'User-Agent': 'PyGithub/Python'}
+None
+200
+[('Server', 'GitHub.com'), ('Date', 'Thu, 01 Feb 2024 23:38:55 GMT'), ('Content-Type', 'application/json; charset=utf-8'), ('Transfer-Encoding', 'chunked'), ('Cache-Control', 'private, max-age=60, s-maxage=60'), ('Vary', 'Accept, Authorization, Cookie, X-GitHub-OTP, Accept-Encoding, Accept, X-Requested-With'), ('ETag', 'W/"fdd3ad04d14a1c6deb0a4a1fbcc104a59889084ffa9e75d80a66f6b7a54ec074"'), ('Last-Modified', 'Wed, 17 Jan 2024 22:24:36 GMT'), ('X-OAuth-Scopes', 'admin:org, admin:public_key, repo, user'), ('X-Accepted-OAuth-Scopes', 'admin:org, read:org, repo, user, write:org'), ('github-authentication-token-expiration', '2024-03-02 22:39:38 UTC'), ('X-GitHub-Media-Type', 'github.v3; format=json'), ('x-github-api-version-selected', '2022-11-28'), ('X-RateLimit-Limit', '5000'), ('X-RateLimit-Remaining', '4994'), ('X-RateLimit-Reset', '1706834317'), ('X-RateLimit-Used', '6'), ('X-RateLimit-Resource', 'core'), ('Access-Control-Expose-Headers', 'ETag, Link, Location, Retry-After, X-GitHub-OTP, X-RateLimit-Limit, X-RateLimit-Remaining, X-RateLimit-Used, X-RateLimit-Resource, X-RateLimit-Reset, X-OAuth-Scopes, X-Accepted-OAuth-Scopes, X-Poll-Interval, X-GitHub-Media-Type, X-GitHub-SSO, X-GitHub-Request-Id, Deprecation, Sunset'), ('Access-Control-Allow-Origin', '*'), ('Strict-Transport-Security', 'max-age=31536000; includeSubdomains; preload'), ('X-Frame-Options', 'deny'), ('X-Content-Type-Options', 'nosniff'), ('X-XSS-Protection', '0'), ('Referrer-Policy', 'origin-when-cross-origin, strict-origin-when-cross-origin'), ('Content-Security-Policy', "default-src 'none'"), ('Content-Encoding', 'gzip'), ('X-GitHub-Request-Id', 'EDD5:2508B2:9AE6C80:9C7B574:65BC2B8F')]
+{"login":"demoorg","id":156956893,"node_id":"O_kgDOCVr43Q","url":"https://api.github.com/orgs/demoorg","repos_url":"https://api.github.com/orgs/demoorg/repos","events_url":"https://api.github.com/orgs/demoorg/events","hooks_url":"https://api.github.com/orgs/demoorg/hooks","issues_url":"https://api.github.com/orgs/demoorg/issues","members_url":"https://api.github.com/orgs/demoorg/members{/member}","public_members_url":"https://api.github.com/orgs/demoorg/public_members{/member}","avatar_url":"https://avatars.githubusercontent.com/u/156956893?v=4","description":null,"is_verified":false,"has_organization_projects":true,"has_repository_projects":true,"public_repos":3,"public_gists":0,"followers":0,"following":0,"html_url":"https://github.com/demoorg","created_at":"2024-01-17T20:14:26Z","updated_at":"2024-01-17T22:24:36Z","archived_at":null,"type":"Organization","total_private_repos":0,"owned_private_repos":0,"private_gists":0,"disk_usage":5,"collaborators":0,"billing_email":"me@example.com","default_repository_permission":"read","members_can_create_repositories":true,"two_factor_requirement_enabled":true,"members_allowed_repository_creation_type":"all","members_can_create_public_repositories":true,"members_can_create_private_repositories":true,"members_can_create_internal_repositories":false,"members_can_create_pages":true,"members_can_fork_private_repositories":false,"web_commit_signoff_required":false,"members_can_create_public_pages":true,"members_can_create_private_pages":true,"plan":{"name":"free","space":976562499,"private_repos":10000,"filled_seats":2,"seats":0},"advanced_security_enabled_for_new_repositories":false,"dependabot_alerts_enabled_for_new_repositories":false,"dependabot_security_updates_enabled_for_new_repositories":false,"dependency_graph_enabled_for_new_repositories":false,"secret_scanning_enabled_for_new_repositories":false,"secret_scanning_push_protection_enabled_for_new_repositories":false,"secret_scanning_push_protection_custom_link_enabled":false,"secret_scanning_push_protection_custom_link":null,"secret_scanning_validity_checks_enabled":false}
